add command line entrypont play_xiangqi_new for app_simple.py
